fix: set desktop menu in a working state
