+++ src/lib/components/mollecules/Header/Header.svelte
@@ -4,6 +4,7 @@
 	import MenuMobile from '../MenuMobile/MenuMobile.svelte';
 	import DesktopMenu from '../DesktopMenu/DesktopMenu.svelte';
 	import SkipLink from '$lib/components/atoms/SkipLink/SkipLink.svelte';
+	import HeaderLogo from './HeaderLogo.svelte';
 
 	export let isMenuOpen = false;
 </script>
@@ -17,53 +18,11 @@
 		role="navigation"
 		aria-label="menu principal"
 	>
-		<a
-			class="text-white lg:hidden xl:block"
-			title="MJ Communication &
-		transition écologique - Accueil"
-			href="/"
-		>
-			<Icon
-				role="img"
-				ariaLabel="MJ Communication &
-			transition écologique"
-				height="auto"
-				width="130px"
-				name="mjConseilFull"
-			/>
-		</a>
-
-		<a class="text-white xl:hidden md:hidden lg:block" aria-label="Vers la page d'accueil" href="/">
-			<Icon
-				role="img"
-				ariaLabel="MJ Communication &
-		transition écologique"
-				height="auto"
-				width="75px"
-				name="mjConseilFull"
-			/>
-		</a>
-
+		<HeaderLogo />
 		<DesktopMenu activeRoute={$page.route.id} />
 	</nav>
 
 	<!-- Menu Mobile -->
-	<div class="flex lg:hidden justify-between w-full">
-		<a class="text-white" aria-label="Vers la page d'accueil" href="/">
-			<Icon height="35px" width="auto" name="mjConseilFull" />
-		</a>
-		<button on:click on:click={() => (isMenuOpen = true)} aria-label="ouvrir le menu de navigation">
-			<Icon height="2rem" width="2rem" name="burger" />
-		</button>
-		<MenuMobile
-			on:clickLink={() => (isMenuOpen = false)}
-			on:clickLink
-			on:click={() => (isMenuOpen = false)}
-			on:closeModal={() => (isMenuOpen = false)}
-			on:closeModal
-			{isMenuOpen}
-		/>
-	</div>
 </div>
 
 <style>
